requirements.txt
ipykernel
tensorflow
pillow
tqdm
opencv-python
matplotlib
scikit-learn
fastapi 
uvicorn[standard]
